refactor

+++ light/client.go
@@ -6,7 +6,7 @@ func NewLightClient() *TwampLightClient {
 	return &TwampLightClient{}
 }
 
-func (c *TwampLightClient) Connect(hostname string, port uint16) (*TwampLightConnection, error) {
+func (c *TwampLightClient) Connect(hostname string, port int) (*TwampLightConnection, error) {
 	twampConnection := NewTwampLightConnection(hostname, port)
 	return twampConnection, nil
 }
